LayoutTests/transitions/transition-end-event-prefixed-02-expected.txt

Test to make sure that if prefixed transition events are modified we correctly modify unprefixed events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS transitionEventBox.currentTarget is transitionEventBox.target
PASS successfullyParsed is true

TEST COMPLETE
